separate token validator; get rid of the big holValidators file again after merge
